reverted back to the original AmazingConverter, given it is actually being used outside.
